Convert physical positioning properties to logical ones where possible
This makes it ready for the layout to be easily converted from LTR to RTL. The logical CSS properties for "left" and "right" are not yet fully supported in some browsers. So this commit adds TODOs for fixing "left" and "right" usages. When major browsers start supporting the logical alternatives (which are inset-inline-start and inset-inline-end), the TODOs could be resolved. After that, the layout will be almost completely ready for applying RTL direction just by adding "direction:rtl" to the rule *{...} when the layout is required to be in RTL direction.